requirements.txt

Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
drf-spectacular>=0.27
python-dotenv>=1.0
PyJWT>=2.8
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9


